fix: Rename conflicting `platform` module
It conflicts with a module name used in Pandas [1]. This is the quick fix.

TOWARDS PLA-209

[1]
```
poetry run python scripts/evaluate.py --wikibase-id Q992
Traceback (most recent call last):
  File "climatepolicyradar/knowledge-graph/scripts/evaluate.py", line 4, in <module>
    import pandas as pd
  File "climatepolicyradar/knowledge-graph/.venv/lib/python3.10/site-packages/pandas/__init__.py", line 26, in <module>
    from pandas.compat import (
  File "climatepolicyradar/knowledge-graph/.venv/lib/python3.10/site-packages/pandas/compat/__init__.py", line 17, in <module>
    from pandas.compat._constants import (
  File "climatepolicyradar/knowledge-graph/.venv/lib/python3.10/site-packages/pandas/compat/_constants.py", line 19, in <module>
    PYPY = platform.python_implementation() == "PyPy"
AttributeError: module 'platform' has no attribute 'python_implementation'
```
