+++ tests/test_train.py
@@ -4,7 +4,7 @@ from unittest.mock import ANY, Mock, patch
 import pytest
 import wandb
 
-from scripts.platform import AwsEnv
+from scripts.cloud import AwsEnv
 from scripts.train import (
     Namespace,
     StorageLink,
